==> code/31_regression.Rmd <==
---
title: "Mine Basin Regressions"
author: "Lukas Vashold"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

```{r prelims, include=FALSE}
library("dplyr")
library("readr")
library("fixest")
library("rdrobust")
sapply(list.files("../R", ".R$"), \(f) {source(paste0("../R/", f)); TRUE})
```

```{r restrictions, include=FALSE}
restr_year <- 2016:2023 # years
restr_area_mined <- 0 # minimum of mined area in mine basin
restr_order <- 10 # maximum order of basins to include (up to 25)
excl_mine_basin <- TRUE # should the mine basin itself be excluded?
mine_downstream <- TRUE # if included, should the mine basin downstream?
restr_number_basins <- 0 # minimum number of up/downstream basins each mine basin has to have

type <- "centroid" # c("", "ipis", "centroid") cent with restr_order = 10 to get pre-revision sample
order <- 10

# Different EVI/NDVI masks and computations
v_mask <- "cci_veg_broad" # c("nomask", "cci_veg_broad", "cci_veg_narrow", "esri_veg")
c_mask <- "cci_c_broad" # c("cci_c_broad", "cci_c_narrow", "cci_c_rainfed", "cci_c_irrigated", "af_c", "esri_c")

comp_max <- "16" # c("16", "px")

measure <- "EVI" # c("EVI", "NDVI")

spec_general_max <- paste0("max_", measure, "_", comp_max, "_", v_mask)
spec_croplands_max <- paste0("max_", measure, "_", comp_max, "_", c_mask)

spec_general_mean <- paste0("mean_", measure, "_16_", v_mask)
spec_croplands_mean <- paste0("mean_", measure, "_16_", c_mask)

# covariates 
covs <- c("elevation", "slope", "soilgrid_grouped", "tmp_max", "precipitation", "accessibility_to_cities_2015", "pop_2015") 
# 

# FEs 
fe <- c("year", "as.factor(mine_basin)")

# cluster SEs
cluster <- "mine_basin"


f_order <- paste0("c(", spec_general_max, ",", spec_croplands_max, ")~", 
                  "i(dist_order, ref = -1)")
f_dist <- paste0("c(", spec_general_max, ",", spec_croplands_max, ")~", 
                 "(dist_km + I(dist_km^2)) * downstream")
f_covs <- paste0("+", paste0(covs, collapse = "+"))
f_fes <- paste0("|", paste0(fe, collapse = "+"))



# load data
df_reg <- readRDS(p("processed/df_reg", 
                         "-order-", order,
                         ifelse(type == "", ".RDS", paste0("-", type, ".RDS"))))

mine_size_restr <- df_reg |> filter(mine_area_km2 > restr_area_mined) |> pull(mine_basin) |> unique()

df_reg_restr <- df_reg |> 
  filter(dist_n <= restr_order, 
         year %in% restr_year, 
         mine_basin %in% mine_size_restr, 
         if(excl_mine_basin) dist_n > 0 else dist_n >= 0)
if(!mine_downstream) {
  df_reg_restr <- df_reg_restr |> 
    mutate(downstream = replace(downstream, dist_n == 0, 0))
}
if(restr_number_basins > 0) {
  mine_number_restr <- df_reg |> filter(year == restr_year[1], dist_n != 0) |> 
  group_by(mine_basin, downstream, dist_n) |> count() |> 
  group_by(mine_basin, downstream) |> count() |> 
  filter(n >= restr_number_basins) |> 
  group_by(mine_basin) |> count() |> 
  filter(n == 2) |> 
  pull(mine_basin)
  df_reg_restr <- df_reg_restr |> 
    filter(mine_basin %in% mine_number_restr)
}
# df_reg_restr <- df_reg_restr |>
#   mutate(order_new = ifelse(downstream == 0, -order, order))
```


## Description

This Markdown contains preliminary results for RDD regressions of mines and their effects on agricultural productivity. The structure of the document is as follows:

* Main models
  - Interaction of downstream indicator with order factor
  - Interaction of downstream indicator with squared distance measure

* Robustness
  - Controls sequentially added
  - Different cropland masks
  - Different max computations
  - Mean EVI instead of Max EVI
  - Varying fixed effects (higher order basins, countries, regions)
  - At least one up-/downstream basin
  - Maximum order of 1
  
* Heterogeneity
  - By biome (forests, deserts, grasslands)
  - By region (North & East Africa, West Africa, Southern Africa)
  - By primary crop (to-do)
  - By primary commodity
  - By mine size
  - By mine activity
  
This specification includes:

* years from **`r min(df_reg_restr$year)` to `r max(df_reg_restr$year)`**
* basin systems with area mined of at least **`r restr_area_mined` square km**
* up/downstream basins up to **`r max(df_reg_restr$dist_n)`** basins "away"
* the mine basin is  **`r ifelse(excl_mine_basin, "not included", ifelse(mine_downstream, "included as downstream", "included as upstream"))`**
* mine basins that have at least **`r restr_number_basins`** up/downstream basin(s)


## Main Models

### Order (factor) interaction

```{r reg_main_order_restr, warning=FALSE, message=FALSE}
# no covariates, order interaction
mod1_order_restr = feols(as.formula(paste0(f_order, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

# with covariates, order interaction
mod2_order_restr = feols(as.formula(paste0(f_order, f_covs, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

etable(mod1_order_restr[1], mod2_order_restr[1],
       mod1_order_restr[2], mod2_order_restr[2],
       drop = "soilgrid")

evi_v_avg <- df_reg_restr |>
  filter(!is.na(get(spec_general_max))) |>
  pull(get(spec_general_max)) |> mean()
evi_c_avg <- df_reg_restr |>
  filter(!is.na(get(spec_croplands_max))) |>
  pull(get(spec_croplands_max)) |> mean()

iplot(mod1_order_restr, ci_level = 0.95,
         main = "Effects by order of basins (without controls)", 
      sub = "Black = Vegetation EVI, Red = Croplands EVI", 
         dict = c("dist_order" = "Order"))
iplot(mod2_order_restr, ci_level = 0.95,
         main = "Effects by order of basins (with controls)", 
      sub = "Black = Vegetation EVI, Red = Croplands EVI", 
         dict = c("dist_order" = "Order"))
```

Compared to the sample mean, the presence of a mine upstream changes

* the vegetation `r measure` (mask: `r v_mask`, max comp: `r comp_max`) by **`r round(coef(mod1_order_restr[[1]])["dist_order::1"] / evi_v_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_restr[[1]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_restr[[1]])["dist_order::1"] / evi_v_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_restr[[1]])["dist_order::1"], 3)`*)
* the croplands `r measure` (mask: `r c_mask`, max comp: `r comp_max`) by **`r round(coef(mod1_order_restr[[2]])["dist_order::1"] / evi_c_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_restr[[2]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_restr[[2]])["dist_order::1"] / evi_c_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_restr[[2]])["dist_order::1"], 3)`*)

on average **without** (with) controls.


### Distance (squared) interaction

```{r reg_main_dist_restr, warning=FALSE, message=FALSE}
# no covariates, squared distance
mod1_dist_restr = feols(as.formula(paste0(f_dist, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)


# with covariates, squared distance
mod2_dist_restr = feols(as.formula(paste0(f_dist, f_covs, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

etable(mod1_dist_restr[1], mod2_dist_restr[1],
       mod1_dist_restr[2], mod2_dist_restr[2],
       drop = "soilgrid")

evi_v_avg <- df_reg_restr |>
  filter(!is.na(get(spec_general_max))) |>
  pull(get(spec_general_max)) |> mean()
evi_c_avg <- df_reg_restr |>
  filter(!is.na(get(spec_croplands_max))) |>
  pull(get(spec_croplands_max)) |> mean()
```

Compared to the sample mean, the presence of a mine upstream changes

* the vegetation `r measure` (mask: `r v_mask`, max comp: `r comp_max`) by **`r round(coef(mod1_dist_restr[[1]])["downstream"] / evi_v_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_dist_restr[[1]])["downstream"], 3)`* (`r round(coef(mod2_dist_restr[[1]])["downstream"] / evi_v_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_dist_restr[[1]])["downstream"], 3)`*)
* the croplands `r measure` (mask: `r c_mask`, max comp: `r comp_max`) by **`r round(coef(mod1_dist_restr[[2]])["downstream"] / evi_c_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_dist_restr[[2]])["downstream"], 3)`* (`r round(coef(mod2_dist_restr[[2]])["downstream"] / evi_c_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_dist_restr[[2]])["downstream"], 3)`*)

on average **without** (with) controls.


## Robustness

### Controls sequentially added

```{r reg_order_rob_addcontr, warning=FALSE, message=FALSE}
# no covariates
mod1_order_contr = feols(as.formula(paste0(f_order, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

# with geo covariates
mod2_order_contr = feols(as.formula(paste0(f_order, "+elevation+slope+soilgrid_grouped", f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

# with geo + met (CRU) covariates
mod3_order_contr = feols(as.formula(paste0(f_order, "+elevation+slope+soilgrid_grouped+tmp_max+precipitation", f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

# with geo + met (TC+CHIRPS) covariates
mod4_order_contr = feols(as.formula(paste0(f_order, "+elevation+slope+soilgrid_grouped+tmax_tc+pre_chirps", f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

iplot(mod1_order_contr, ci_level = 0.95,
         main = "Effects by order of basins (without controls)", 
      sub = "Black = Vegetation EVI, Red = Cropland EVI", 
         dict = c("dist_order" = "Order"))

iplot(mod2_order_contr, ci_level = 0.95,
         main = "Effects by order of basins (with geo controls)", 
      sub = "Black = Vegetation EVI, Red = Cropland EVI", 
         dict = c("dist_order" = "Order"))

iplot(mod3_order_contr, ci_level = 0.95,
         main = "Effects by order of basins (with geo + met (CRU) controls)", 
      sub = "Black = Vegetation EVI, Red = Cropland EVI", 
         dict = c("dist_order" = "Order"))

iplot(mod4_order_contr, ci_level = 0.95,
         main = "Effects by order of basins (with geo + met (TC+CHIRPS) controls)", 
      sub = "Black = Vegetation EVI, Red = Cropland EVI", 
         dict = c("dist_order" = "Order"))
```


### Different measure (NDVI)

```{r reg_order_rob_measures, warning=FALSE, message=FALSE}
# EVI/NDVI measures
measures <- c("EVI", "NDVI")

specs_general_max <- paste0("max_", measures, "_", comp_max, "_", v_mask)
specs_croplands_max <- paste0("max_", measures, "_", comp_max, "_", c_mask)

f_order_measures_veg <- paste0("c(", paste0(c(specs_general_max), collapse = ","), ")~", 
                            "i(dist_order, ref = -1)")

f_order_measures_c <- paste0("c(", paste0(c(specs_croplands_max), collapse = ","), ")~", 
                            "i(dist_order, ref = -1)")

# no covariates
mod1_order_measures_veg = feols(as.formula(paste0(f_order_measures_veg, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)
# with covariates
mod2_order_measures_veg = feols(as.formula(paste0(f_order_measures_veg, f_covs, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

# no covariates
mod1_order_measures_c = feols(as.formula(paste0(f_order_measures_c, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)
# with covariates
mod2_order_measures_c = feols(as.formula(paste0(f_order_measures_c, f_covs, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

evi_v_EVI_avg <- df_reg_restr |>
  filter(!is.na(get(specs_general_max[which(grepl("EVI", specs_general_max))]))) |>
  pull(get(specs_general_max[which(grepl("EVI", specs_general_max))])) |> mean()
evi_c_EVI_avg <- df_reg_restr |>
  filter(!is.na(get(specs_croplands_max[which(grepl("EVI", specs_croplands_max))]))) |>
  pull(get(specs_croplands_max[which(grepl("EVI", specs_croplands_max))])) |> mean()

evi_v_NDVI_avg <- df_reg_restr |>
  filter(!is.na(get(specs_general_max[which(grepl("NDVI", specs_general_max))]))) |>
  pull(get(specs_general_max[which(grepl("NDVI", specs_general_max))])) |> mean()
evi_c_NDVI_avg <- df_reg_restr |>
  filter(!is.na(get(specs_croplands_max[which(grepl("NDVI", specs_croplands_max))]))) |>
  pull(get(specs_croplands_max[which(grepl("NDVI", specs_croplands_max))])) |> mean()


# iplot(mod1_order_measures_veg, ci_level = 0.95, col = seq_len(length(measures)),
#       main = "Vegetation NDVI vs EVI (without controls)", 
#       dict = c("dist_order" = "Order"))
# legend("bottomleft", col = seq_len(length(measures)), pch = 20,
#        legend = measures, title = "Max Computation")
# 
# iplot(mod1_order_measures_c, ci_level = 0.95, col = seq_len(length(measures)),
#       main = "Croplands NDVI vs EVI (without controls)", 
#       dict = c("dist_order" = "Order"))
# legend("bottomleft", col = seq_len(length(measures)), pch = 20, lwd = 1,
#        legend = measures, title = "Max Computation")

iplot(mod2_order_measures_veg, ci_level = 0.95, col = seq_len(length(measures)),
      main = "Vegetation NDVI vs EVI (with controls)",
      dict = c("dist_order" = "Order"))
legend("bottomleft", col = seq_len(length(measures)), pch = 20,
       legend = measures, title = "Measure")

iplot(mod2_order_measures_c, ci_level = 0.95, col = seq_len(length(measures)),
      main = "Croplands NDVI vs EVI (with controls)",
      dict = c("dist_order" = "Order"))
legend("bottomleft", col = seq_len(length(measures)), pch = 20, lwd = 1,
       legend = measures, title = "Measure")
```

Compared to the sample mean, the presence of a mine upstream changes

* the vegetation `r measures[1]` (mask: `r v_mask`, max comp: `r comp_max`) by **`r round(coef(mod1_order_measures_veg[[1]])["dist_order::1"] / evi_v_EVI_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_measures_veg[[1]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_measures_veg[[1]])["dist_order::1"] / evi_v_EVI_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_measures_veg[[1]])["dist_order::1"], 3)`*)
* the vegetation `r measures[2]` (mask: `r v_mask`, max comp: `r comp_max`) by **`r round(coef(mod1_order_measures_veg[[2]])["dist_order::1"] / evi_v_NDVI_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_measures_veg[[2]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_measures_veg[[2]])["dist_order::1"] / evi_v_NDVI_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_measures_veg[[2]])["dist_order::1"], 3)`*)

* the croplands `r measures[1]` (mask: `r c_mask`, max comp: `r comp_max`) by **`r round(coef(mod1_order_measures_c[[1]])["dist_order::1"] / evi_c_EVI_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_measures_c[[1]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_measures_c[[1]])["dist_order::1"] / evi_c_EVI_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_measures_c[[1]])["dist_order::1"], 3)`*)
* the croplands `r measures[2]` (mask: `r c_mask`, max comp: `r comp_max`) by **`r round(coef(mod1_order_measures_c[[2]])["dist_order::1"] / evi_c_NDVI_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_measures_c[[2]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_measures_c[[2]])["dist_order::1"] / evi_c_NDVI_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_measures_c[[2]])["dist_order::1"], 3)`*)

on average **without** (with) controls.


### Different vegetation/cropland masks

```{r reg_order_rob_masks, warning=FALSE, message=FALSE}
# Different EVI/NDVI masks
v_masks <- c("cci_veg_broad", "cci_veg_narrow") # , "esri_veg"
c_masks <- c("cci_c_broad", "cci_c_narrow", "cci_c_rainfed", "cci_c_irrigated", "af_c") # , "esri_c"

specs_general_max <- paste0("max_", measure, "_", comp_max, "_", v_masks)
specs_croplands_max <- paste0("max_", measure, "_", comp_max, "_", c_masks)

f_order_masks_veg <- paste0("c(", paste0(c(specs_general_max), collapse = ","), ")~", 
                            "i(dist_order, ref = -1)")

f_order_masks_c <- paste0("c(", paste0(c(specs_croplands_max), collapse = ","), ")~", 
                            "i(dist_order, ref = -1)")

# no covariates
mod1_order_masks_veg = feols(as.formula(paste0(f_order_masks_veg, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)
# with covariates
mod2_order_masks_veg = feols(as.formula(paste0(f_order_masks_veg, f_covs, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

# no covariates
mod1_order_masks_c = feols(as.formula(paste0(f_order_masks_c, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)
# with covariates
mod2_order_masks_c = feols(as.formula(paste0(f_order_masks_c, f_covs, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

# iplot(mod1_order_masks_veg, ci_level = 0.95, col = seq_len(length(v_masks)),
#          main = "Effects by order of basins (without controls)", 
#          dict = c("dist_order" = "Order"))
# legend("bottomleft", col = seq_len(length(v_masks)), pch = 20,
#        legend = v_masks, title = "Vegetation Mask")
# 
# iplot(mod1_order_masks_c, ci_level = 0.95, col = seq_len(length(c_masks)),
#          main = "Effects by order of basins (without controls)", 
#          dict = c("dist_order" = "Order"))
# legend("bottomleft", col = seq_len(length(c_masks)), pch = 20, lwd = 1,
#        legend = c_masks, title = "Cropland Mask")

iplot(mod2_order_masks_veg, ci_level = 0.95, col = seq_len(length(v_masks)),
         main = "Different vegetation masks (with controls)", 
         dict = c("dist_order" = "Order"))
legend("bottomleft", col = seq_len(length(v_masks)), pch = 20,
       legend = v_masks, title = "Vegetation Mask")

iplot(mod2_order_masks_c, ci_level = 0.95, col = seq_len(length(c_masks)),
         main = "Different cropland masks (with controls)", 
         dict = c("dist_order" = "Order"))
legend("bottomleft", col = seq_len(length(c_masks)), pch = 20, lwd = 1,
       legend = c_masks, title = "Cropland Mask")

```


### Different Max calculation

```{r reg_order_rob_comps, warning=FALSE, message=FALSE}
# Different max computations
comp_maxs <- c("16", "px")

specs_general_max <- paste0("max_", measure, "_", comp_maxs, "_", v_mask)
specs_croplands_max <- paste0("max_", measure, "_", comp_maxs, "_", c_mask)

f_order_comps_veg <- paste0("c(", paste0(c(specs_general_max), collapse = ","), ")~", 
                            "i(dist_order, ref = -1)")

f_order_comps_c <- paste0("c(", paste0(c(specs_croplands_max), collapse = ","), ")~", 
                            "i(dist_order, ref = -1)")

# no covariates
mod1_order_comps_veg = feols(as.formula(paste0(f_order_comps_veg, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)
# with covariates
mod2_order_comps_veg = feols(as.formula(paste0(f_order_comps_veg, f_covs, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

# no covariates
mod1_order_comps_c = feols(as.formula(paste0(f_order_comps_c, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)
# with covariates
mod2_order_comps_c = feols(as.formula(paste0(f_order_comps_c, f_covs, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

evi_v_16_avg <- df_reg_restr |>
  filter(!is.na(get(specs_general_max[which(grepl("16", specs_general_max))]))) |>
  pull(get(specs_general_max[which(grepl("16", specs_general_max))])) |> mean()
evi_c_16_avg <- df_reg_restr |>
  filter(!is.na(get(specs_croplands_max[which(grepl("16", specs_croplands_max))]))) |>
  pull(get(specs_croplands_max[which(grepl("16", specs_croplands_max))])) |> mean()

evi_v_px_avg <- df_reg_restr |>
  filter(!is.na(get(specs_general_max[which(grepl("px", specs_general_max))]))) |>
  pull(get(specs_general_max[which(grepl("px", specs_general_max))])) |> mean()
evi_c_px_avg <- df_reg_restr |>
  filter(!is.na(get(specs_croplands_max[which(grepl("px", specs_croplands_max))]))) |>
  pull(get(specs_croplands_max[which(grepl("px", specs_croplands_max))])) |> mean()


# iplot(mod1_order_comps_veg, ci_level = 0.95, col = seq_len(length(comp_maxs)),
#       main = "Effects by order of basins (without controls)", 
#       sub = paste0("Vegetation ", measure),
#       dict = c("dist_order" = "Order"))
# legend("bottomleft", col = seq_len(length(comp_maxs)), pch = 20,
#        legend = comp_maxs, title = "Max Computation")
# 
# iplot(mod1_order_comps_c, ci_level = 0.95, col = seq_len(length(comp_maxs)),
#       main = "Effects by order of basins (without controls)", 
#       sub = paste0("Croplands ", measure),
#       dict = c("dist_order" = "Order"))
# legend("bottomleft", col = seq_len(length(comp_maxs)), pch = 20, lwd = 1,
#        legend = comp_maxs, title = "Max Computation")

iplot(mod2_order_comps_veg, ci_level = 0.95, col = seq_len(length(comp_maxs)),
      main = "Effects by order of basins (with controls)", 
      sub = paste0("Vegetation ", measure),
      dict = c("dist_order" = "Order"))
legend("bottomleft", col = seq_len(length(comp_maxs)), pch = 20,
       legend = comp_maxs, title = "Max Computation")

iplot(mod2_order_comps_c, ci_level = 0.95, col = seq_len(length(comp_maxs)),
      main = "Effects by order of basins (with controls)", 
      sub = paste0("Croplands ", measure),
      dict = c("dist_order" = "Order"))
legend("bottomleft", col = seq_len(length(comp_maxs)), pch = 20, lwd = 1,
       legend = comp_maxs, title = "Max Computation")
```

Compared to the sample mean, the presence of a mine upstream changes

* the vegetation `r measure` (mask: `r v_mask`, max comp: `r comp_maxs[1]`) by **`r round(coef(mod1_order_comps_veg[[1]])["dist_order::1"] / evi_v_16_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_comps_veg[[1]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_comps_veg[[1]])["dist_order::1"] / evi_v_16_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_comps_veg[[1]])["dist_order::1"], 3)`*)
* the vegetation `r measure` (mask: `r v_mask`, max comp: `r comp_maxs[2]`) by **`r round(coef(mod1_order_comps_veg[[2]])["dist_order::1"] / evi_v_px_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_comps_veg[[2]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_comps_veg[[2]])["dist_order::1"] / evi_v_px_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_comps_veg[[2]])["dist_order::1"], 3)`*)

* the croplands `r measure` (mask: `r c_mask`, max comp: `r comp_maxs[1]`) by **`r round(coef(mod1_order_comps_c[[1]])["dist_order::1"] / evi_v_16_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_comps_c[[1]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_comps_c[[1]])["dist_order::1"] / evi_v_16_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_comps_c[[1]])["dist_order::1"], 3)`*)
* the croplands `r measure` (mask: `r c_mask`, max comp: `r comp_maxs[2]`) by **`r round(coef(mod1_order_comps_c[[2]])["dist_order::1"] / evi_v_px_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_comps_c[[2]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_comps_c[[2]])["dist_order::1"] / evi_v_px_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_comps_c[[2]])["dist_order::1"], 3)`*)

on average **without** (with) controls.


### Mean instead of Max

```{r reg_order_rob_meanEVI, warning=FALSE, message=FALSE}
f_order_mean <- paste0("c(", spec_general_mean , ",", spec_croplands_mean, ")~", 
                       "i(dist_order, ref = -1)")

# no covariates, order interaction
mod1_order_meanEVI = feols(as.formula(paste0(f_order_mean, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

# with covariates, order interaction
mod2_order_meanEVI = feols(as.formula(paste0(f_order_mean, f_covs, f_fes)),
                         data = df_reg_restr,
                         cluster = cluster)

evi_v_avg <- df_reg_restr |>
  filter(!is.na(get(spec_general_mean))) |>
  pull(get(spec_general_mean)) |> mean()
evi_c_avg <- df_reg_restr |>
  filter(!is.na(get(spec_croplands_mean))) |>
  pull(get(spec_croplands_mean)) |> mean()

# iplot(mod1_order_meanEVI, ci_level = 0.95,
#          main = "Effects by order of basins (without controls)", 
#       sub = "Black = Vegetation EVI, Red = Croplands EVI", 
#          dict = c("dist_order" = "Order"))
iplot(mod2_order_meanEVI, ci_level = 0.95,
         main = "Mean EVI instead of Max (with controls)", 
      sub = "Black = Vegetation EVI, Red = Croplands EVI", 
         dict = c("dist_order" = "Order"))
```

Compared to the sample mean, the presence of a mine upstream changes

* the vegetation `r measure` (mask: `r v_mask`, mean) by **`r round(coef(mod1_order_meanEVI[[1]])["dist_order::1"] / evi_v_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_meanEVI[[1]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_meanEVI[[1]])["dist_order::1"] / evi_v_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_meanEVI[[1]])["dist_order::1"], 3)`*)
* the croplands `r measure` (mask: `r c_mask`, mean) by **`r round(coef(mod1_order_meanEVI[[2]])["dist_order::1"] / evi_c_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_meanEVI[[2]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_meanEVI[[2]])["dist_order::1"] / evi_c_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_meanEVI[[2]])["dist_order::1"], 3)`*)

on average **without** (with) controls.


### At least one up-/downstream basin

```{r reg_order_atleast1, warning=FALSE, message=FALSE}
mine_number_restr <- df_reg_restr |> filter(year == restr_year[1], dist_n != 0) |> 
  group_by(mine_basin, downstream, dist_n) |> count() |> 
  group_by(mine_basin, downstream) |> count() |> 
  filter(n >= 1) |> 
  group_by(mine_basin) |> count() |> 
  filter(n == 2) |> 
  pull(mine_basin)
df_reg_atleast1 <- df_reg_restr |> 
  filter(mine_basin %in% mine_number_restr)

# no covariates, order interaction
mod1_order_atleast1 = feols(as.formula(paste0(f_order, f_fes)),
                            data = df_reg_atleast1,
                            cluster = cluster)

# with covariates, order interaction
mod2_order_atleast1 = feols(as.formula(paste0(f_order, f_covs, f_fes)),
                            data = df_reg_atleast1,
                            cluster = cluster)

evi_v_avg <- df_reg_atleast1 |>
  filter(!is.na(get(spec_general_max))) |>
  pull(get(spec_general_max)) |> mean()
evi_c_avg <- df_reg_atleast1 |>
  filter(!is.na(get(spec_croplands_max))) |>
  pull(get(spec_croplands_max)) |> mean()

# iplot(mod1_order_atleast1, ci_level = 0.95,
#          main = "Effects by order of basins (without controls)", 
#       sub = "Black = Vegetation EVI, Red = Croplands EVI", 
#          dict = c("dist_order" = "Order"))
iplot(mod2_order_atleast1, ci_level = 0.95,
         main = "At least one up/downstream (with controls)", 
      sub = "Black = Vegetation EVI, Red = Croplands EVI", 
         dict = c("dist_order" = "Order"))
```

Compared to the sample mean, the presence of a mine upstream changes

* the vegetation `r measure` (mask: `r v_mask`, max comp: `r comp_max`) by **`r round(coef(mod1_order_atleast1[[1]])["dist_order::1"] / evi_v_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_atleast1[[1]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_atleast1[[1]])["dist_order::1"] / evi_v_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_atleast1[[1]])["dist_order::1"], 3)`*)
* the croplands `r measure` (mask: `r c_mask`, max comp: `r comp_max`) by **`r round(coef(mod1_order_atleast1[[2]])["dist_order::1"] / evi_c_avg * 100, 2)`%**, p-value = *`r round(pvalue(mod1_order_atleast1[[2]])["dist_order::1"], 3)`* (`r round(coef(mod2_order_atleast1[[2]])["dist_order::1"] / evi_c_avg * 100, 2)`%, p-value = *`r round(pvalue(mod2_order_atleast1[[2]])["dist_order::1"], 3)`*)

on average **without** (with) controls.



## Heterogeneity

### By biome

```{r reg_order_heterogeneity_biome, warning=FALSE, message=FALSE, include=TRUE}
mod_order_hetero_biome = feols(as.formula(paste0(f_order, f_covs, f_fes)),
                         data = df_reg_restr,
                         split = "biome_group",
                         cluster = cluster)

pos_veg <- which(grepl("_veg_", names(mod_order_hetero_biome)))
pos_c <- which(grepl("_c_", names(mod_order_hetero_biome)))

biome <- c()
for(i in pos_veg) {
  biome_model <- mod_order_hetero_biome[[i]]$model_info$sample$value
  biome <- c(biome, biome_model)
}

iplot(mod_order_hetero_biome[pos_veg], ci_level = 0.95, 
      col = seq_len(length(biome)),
      main = paste0("Vegetation ", measure, " (with controls)"), 
      dict = c("dist_order" = "Order"))
legend("bottomleft", col = seq_len(length(biome)), pch = 20,
       legend = biome, title = "Biome")

iplot(mod_order_hetero_biome[pos_c], ci_level = 0.95, col = seq_len(length(biome)),
      main = paste0("Croplands ", measure, " (with controls)"), 
      dict = c("dist_order" = "Order"))
legend("bottomleft", col = seq_len(length(biome)), pch = 20,
       legend = biome, title = "Biome")
```


### By region

```{r reg_order_heterogeneity_region, warning=FALSE, message=FALSE, include=TRUE}
mod_order_hetero_region = feols(as.formula(paste0(f_order, f_covs, f_fes)),
                                data = df_reg_restr,
                                split = "region_grouped",
                                cluster = cluster)

pos_veg <- which(grepl("_veg_", names(mod_order_hetero_region)))
pos_c <- which(grepl("_c_", names(mod_order_hetero_region)))

region <- c()
for(i in pos_veg) {
  region_model <- mod_order_hetero_region[[i]]$model_info$sample$value
  region <- c(region, region_model)
}

iplot(mod_order_hetero_region[pos_veg], ci_level = 0.95, 
      col = seq_len(length(region)),
      main = paste0("Vegetation ", measure, " (with controls)"), 
      dict = c("dist_order" = "Order"))
legend("bottomleft", col = seq_len(length(region)), pch = 20,
       legend = region, title = "Region")

iplot(mod_order_hetero_region[pos_c], ci_level = 0.95, col = seq_len(length(region)),
      main = paste0("Croplands ", measure, " (with controls)"), 
      dict = c("dist_order" = "Order"))
legend("bottomleft", col = seq_len(length(region)), pch = 20,
       legend = region, title = "Region")
```


### By primary crop

<!-- To-do -->


### By primary commodity

<!-- To-do -->


### By mine size

<!-- To-do -->


### By mine activity

<!-- To-do -->
